Equality operator, simple index tests
I have added support for the equality operator.  It tests if two ropes
are structurally equivalent.  This tests if the node structure is
equivalent, and if the data in the leaf nodes match.  It ignores the
data in the non-leaf nodes.

I have also added some basic unit tests for indexing, currently one the
single-node case is tested.