Index reference bugfix + tests
There were some bugs in the index reads, mostly around the use of modulo
arithmetic.  The new implementation is a bit closer to the CPython index
checking.

We now explicitly bump up the index by one length if it's negative.  We
then check if the normalised index is within bounds, and raise an
exception within the `Rope` (rather than the data).  Finally, with all
the complicated alternative removed, we can write a much simpler
expression for iterating through the tree and returning the character.

Slicing is still utterly broken.