List initialisation + tests
A rope can now be pre-constructed using a list of words.  The tree is
balanced with respect to the words, but not the length.

We divide the list into two, rounding up so that left-arithmetic
associativity is preferred.  That is, `Rope(['a', 'b', 'c'])` returns
`(Rope('a') + Rope('b')) +  Rope('c')` and not
`Rope('a') + (Rope('b') + Rope('c'))`.

This is so we can reproduce the concatenation result without
parenthesis.  That is,
`Rope(['a', 'b', 'c']) == Rope('a') + Rope('b') + Rope('c')`.

Some basic list initialistion tests have also been added.