Word iteration + tests
Iteration over the rope will now output the words in each leaf.

There isn't a lot of control here, nor is there any way to iterate over
characters (which is the expected behaviour of `iter()` on a `str`).

This iterator also should probably be moved to a separate class, and
saved as a property within the rope.