Strides with steps
Strides with steps are now supported.  Not tested exhaustively, but all
the weird cases appear to work (one-node, three-node, negative strides,
etc.)

The logic around getting the start and stop indices is now much more
complicated, because of the flipped logic that occurs when step size is
negative.  This was mostly ripped out from CPython's strideobject.c
code.  It's probably not well-optimised for Python, since it is just a
transliteration of the C source.  Might be worth looking into in the
future (or just rewriting Rope in C...).

Single-node tests were added, but multinode tests are needed.