Rope reference copy on slice, bug fixes
Main difference here is that we now correctly return a copy of the rope
on a slice, even if the data of the original and copy are identical.

I also fixed a few runtime bugs related to negative slicing.

However, negative slicing is still broken.

Testing was also generalised a bit further, which exposed some of these
bugs.  (Testing is also heaps slower now, approaching 1 second.)