Streamling index getitem call
Some of the logic around indexing has been streamlined for direct
indexing of ropes, mainly to eliminate the hard index bound checks and
rely more on the subropes to do those checks.

Mostly this is achieved by converting the global index into the
equivalent subrope index.  This is (sloppily) documented in the
index.txt file.

This is not yet implemented in the slices, but it should work in
principle, and should similarly let us remove all of those bound checks.