Partial implementation of new slicing
This is a partial implementation of the slicing based on the technique
used in indexing (spoofing the global index into the local index).

It only works for slices that are exclusive to individual subropes, and
falls back to the old code when a stride uses both subropes.

This also does not work if there is a negative stride.  Specifically, it
does not realise that it's crossing into a new rope.  (Probably some
index phase issues as well.)