Completed slicing support for new getitem
Slicing support without strides appears to be working now.  The case of
a slice spread across both subropes works.  Also a bug was fixed where a
negative end index on a rope boundary was hard-coded to return None for
the left rope.  This is a bit arbitrary but it may be necessary.

Striding, particularly negative strides, still does not work, so this is
only a partial victory.