new step > 0 slice method; slice test cleanup; len
The new slice method now seems to be working correctly for all positive
step sizes, including a correction to offset calculation and fixing some
new bugs related to (step > 0) slices which start on the right and end
in the left.

Slice testing was also greately cleaned up and expanded to catch the
bugfix case described above (which was, we note, already part of the
stride test for step = 1).

A len() operator was also added which allows us to do a `if rope` style
test.